+++ 14 - APIs com Nodejs e Express/nodejs-buscas-filtros-paginacao-erros-api/src/controllers/livroController.js
@@ -1,8 +1,10 @@
 import livroModel from '../models/Livro.js'
 import { autor as autorModel } from '../models/Autor.js';
+import ValidationError from '../errors/ValidationError.js';
+import NotFoundError from '../errors/NotFoundError.js';
 
 class LivroController {
-    static async listarLivros(req, res) {
+    static async listarLivros(req, res, next) {
         try {
             const listaLivros = await livroModel.find({});
 
@@ -13,27 +15,27 @@ class LivroController {
                 },
             });
         } catch (error) {
-            console.error('Erro ao listar livros', error);
-            res.status(500).json({ error: 'Erro ao listar livros' });
+            next(error);
         }
     }
 
-    static async buscarLivro(req, res) {
+    static async buscarLivro(req, res, next) {
         try {
             const { id } = req.params;
 
+            if (!id.match(/^[0-9a-fA-F]{24}$/)) return new ValidationError('ID inválido', 400).sendResponse(res);
+
             const livro = await livroModel.findById(id);
 
-            if (!livro) return res.status(404).json({ error: 'Livro não encontrado' });
+            if (!livro) return new NotFoundError('Livro não encontrado').sendResponse(res);
 
             res.status(200).json({ data: livro });
         } catch (error) {
-            console.error('Erro ao buscar livro', error);
-            res.status(500).json({ error: 'Erro ao buscar livro' });
+            next(error);
         }
     }
 
-    static async listarLivrosPorEditora(req, res) {
+    static async listarLivrosPorEditora(req, res, next) {
         try {
             const editora = req.query.editora;
 
@@ -46,18 +48,21 @@ class LivroController {
                 },
             });
         } catch (error) {
-            console.error('Erro ao listar livros por editora', error);
-            res.status(500).json({ error: 'Erro ao listar livros por editora' });
+            next(error);
         }
     }
 
-    static async criarLivro(req, res) {
+    static async criarLivro(req, res, next) {
         try {
             const { titulo, editora, valor, paginas, autorId } = req.body;
 
+            if (!autorId) return new ValidationError('ID do autor é obrigatório', 400).sendResponse
+
+            if (!autorId.match(/^[0-9a-fA-F]{24}$/)) return new ValidationError('ID do autor inválido', 400).sendResponse(res);
+
             const autor = await autorModel.findById(autorId);
 
-            if (!autor) return res.status(404).json({ error: 'Autor informado não encontrado' });
+            if (!autor) return new NotFoundError('Autor informado não encontrado').sendResponse(res);
 
             const novoLivro = new livroModel({
                 titulo,
@@ -71,22 +76,26 @@ class LivroController {
 
             res.status(201).json({ message: 'Livro criado com sucesso', data: novoLivro });
         } catch (error) {
-            console.error('Erro ao criar livro', error);
-            res.status(500).json({ error: 'Erro ao criar livro' });
+            next(error);
         }
     }
 
-    static async atualizarLivro(req, res) {
+    static async atualizarLivro(req, res, next) {
         try {
             const { id } = req.params;
+
+            if (!id.match(/^[0-9a-fA-F]{24}$/)) return new ValidationError('ID inválido', 400).sendResponse(res);
+
             const { titulo, editora, valor, paginas, autorId } = req.body;
+
+            if (autorId && !autorId.match(/^[0-9a-fA-F]{24}$/)) return new ValidationError('ID do autor inválido', 400).sendResponse(res);
 
             let autor = null;
 
             if (autorId) {
                 autor = await autorModel.findById(autorId);
 
-                if (!autor) return res.status(404).json({ error: 'Autor informado não encontrado' });
+                if (!autor) return new NotFoundError('Autor informado não encontrado').sendResponse(res);
             }
 
             await livroModel.findByIdAndUpdate(id, {
@@ -99,21 +108,23 @@ class LivroController {
 
             res.status(200).json({ message: 'Livro atualizado com sucesso' });
         } catch (error) {
-            console.error('Erro ao atualizar livro', error);
-            res.status(500).json({ error: 'Erro ao atualizar livro' });
+            next(error);
         }
     }
 
-    static async deletarLivro(req, res) {
+    static async deletarLivro(req, res, next) {
         try {
             const { id } = req.params;
 
-            await livroModel.findByIdAndDelete(id);
+            if (!id.match(/^[0-9a-fA-F]{24}$/)) return new ValidationError('ID inválido', 400).sendResponse(res);
+
+            const livro = await livroModel.findByIdAndDelete(id);
+
+            if (!livro) return new NotFoundError('Livro não encontrado').sendResponse(res);
 
             res.status(200).json({ message: 'Livro deletado com sucesso' });
         } catch (error) {
-            console.error('Erro ao deletar livro', error);
-            res.status(500).json({ error: 'Erro ao deletar livro' });
+            next(error);
         }
     }
 }
